fix: Forward timeout, I2C logging, type truncation, switch default, README pins/semantics
Agent-Logs-Url: https://github.com/wruih139/h750/sessions/e48e2957-7566-4f67-be86-b5e85f1d0733

+++ down_car/trackline.h
@@ -4,4 +4,4 @@
 void Trackline_Init(void);
 void Sensor_Receive(void);
 void Tracking_Line_Task(void);
-void Forward(uint8_t n);
+void Forward(uint16_t n);
